requirements.txt
requests
pytest==5.2.3
allure-pytest>=2.8.6
jsonschema==3.0.1
psycopg2-binary
pathlib==1.0.1
setuptools
jira
cachetools
pytest-xdist==1.30.0
pytest-timeouts==1.2
dateutils
redis
addict
property-cached
slackclient==2.3.1
portainer-api==1.22.1.1
jenkins-job-builder==3.1.0
marshmallow
loguru==0.4.0
sshtunnel==0.1.5
influxdb
kafka-python==1.4.6
